refactor all into smaller parts
