+++ src/client/hooks/tetris-game/constants.js
@@ -1,0 +1,32 @@
+/*
+ ** GRID
+ */
+
+export const FREE = 0;
+export const CURRENT_PIECE = 1;
+
+/*
+ ** SCORING
+ */
+
+export const COMBO = {
+  0: 0,
+  1: 100,
+  2: 300,
+  3: 500,
+  4: 800,
+};
+
+/*
+ ** MOVES
+ */
+
+export const MOVE_LEFT = -1;
+export const MOVE_RIGHT = 1;
+
+/*
+ ** TIMERS
+ */
+
+export const INTERVAL_MS = 750;
+export const DEFAULT_REPEAT_TIMEOUT = 5;
