+++ src/client/hooks/tetris-game/grid/index.js
@@ -1,0 +1,6 @@
+import * as Check from "./checks";
+
+export { Check };
+export { default as write } from "./write";
+export { default as clear } from "./clear";
+export { default as bind } from "./bind";
