+++ src/client/hooks/tetris-game/pieces/index.js
@@ -1,0 +1,4 @@
+export { default as insertion } from "./insertion";
+export { default as lateralMove } from "./lateralMove";
+export { default as rotation } from "./rotation";
+export { default as softDrop } from "./softDrop";
